requirements.txt
# Core LLM
openai>=1.0.0

# Embeddings (local — no API key required)
sentence-transformers>=3.0.0
torch>=2.2.0

# Vector store (local, file-based)
chromadb>=0.5.0

# Evaluation
ragas>=0.2.0
datasets>=2.20.0
langchain-openai>=0.2.0      # RAGAS uses langchain internally as LLM judge wrapper
langchain-text-splitters>=0.3.0  # RecursiveCharacterTextSplitter, token-aware splitters

# Document loading (PDF — use pdfplumber for structured insurance PDFs)
pypdf>=4.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0

# Data manipulation
pandas>=2.1.0
numpy>=1.26.0

# Visualisation
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0           # generate promptfoo evaluation configs
tqdm>=4.66.0
tiktoken>=0.7.0
httpx>=0.27.0
scikit-learn>=1.4.0          # cosine_similarity for embedding analysis
rank-bm25>=0.2.2             # BM25 keyword search (Module 5 hybrid retrieval)

# Notebook
jupyter>=1.0.0
ipywidgets>=8.0.0

# Bonus components
streamlit>=1.35.0           # bonus/03_streamlit_app — chat UI frontend
langdetect>=1.0.9           # bonus/01_data_cleaning — language detection (FR/NL/EN)
